-moved dataset classes to specific modules -added docstrings -formatted example for sphinx -merged preprocessors into raw and window transformers -moved transformers to datautil
